remove some unused deps
