+++ crates/bevy_render/Cargo.toml
@@ -19,10 +19,10 @@ webp = ["image/webp", "bevy_image/webp"]
 dds = ["ddsfile", "bevy_image/dds"]
 pnm = ["image/pnm", "bevy_image/pnm"]
 
-ddsfile = ["dep:ddsfile", "bevy_image/ddsfile"]
+ddsfile = ["bevy_image/ddsfile"]
 ktx2 = ["dep:ktx2", "bevy_image/ktx2"]
-flate2 = ["dep:flate2", "bevy_image/flate2"]
-ruzstd = ["dep:ruzstd", "bevy_image/ruzstd"]
+flate2 = ["bevy_image/flate2"]
+ruzstd = ["bevy_image/ruzstd"]
 basis-universal = ["dep:basis-universal", "bevy_image/basis-universal"]
 
 multi_threaded = ["bevy_tasks/multi_threaded"]
@@ -59,7 +59,6 @@ bevy_ecs = { path = "../bevy_ecs", version = "0.15.0-dev" }
 bevy_encase_derive = { path = "../bevy_encase_derive", version = "0.15.0-dev" }
 bevy_hierarchy = { path = "../bevy_hierarchy", version = "0.15.0-dev" }
 bevy_math = { path = "../bevy_math", version = "0.15.0-dev" }
-bevy_mikktspace = { path = "../bevy_mikktspace", version = "0.15.0-dev" }
 bevy_reflect = { path = "../bevy_reflect", version = "0.15.0-dev", features = [
   "bevy",
 ] }
@@ -91,16 +90,11 @@ wgpu = { version = "22", default-features = false, features = [
 ] }
 naga = { version = "22", features = ["wgsl-in"] }
 serde = { version = "1", features = ["derive"] }
-bitflags = { version = "2.3", features = ["serde"] }
 bytemuck = { version = "1.5", features = ["derive", "must_cast"] }
 downcast-rs = "1.2.0"
 thiserror = "1.0"
 futures-lite = "2.0.1"
-ddsfile = { version = "0.5.2", optional = true }
 ktx2 = { version = "0.3.0", optional = true }
-# For ktx2 supercompression
-flate2 = { version = "1.0.22", optional = true }
-ruzstd = { version = "0.7.0", optional = true }
 # For transcoding of UASTC/ETC1S universal formats, and for .basis file support
 basis-universal = { version = "0.3.0", optional = true }
 encase = { version = "0.10", features = ["glam"] }
